+++ mdl/ast/ast_navigation.go
@@ -86,6 +86,7 @@ func (s *CreateMenuStmt) isStatement() {}
 // DropMenuStmt is `drop menu Module.Name`.
 type DropMenuStmt struct {
 	Name QualifiedName
+	DropIfExists
 }
 
 func (s *DropMenuStmt) isStatement() {}